CPS Artemis: Increase number of rflow lookups in TxDMA (#14001)
- Support upto 6 levels of collisions for rflow in TxDMA for Artemis CPS assist
and fix the bug in hal p4pd that prevented adding those tables this earlier.

+++ nic/sdk/lib/p4/p4pluspd_rxdma_api.cc
@@ -47,7 +47,7 @@ static p4pd_table_properties_t *_p4plus_rxdma_tbls;
 
 namespace pt = boost::property_tree;
 
-static uint8_t
+static int
 p4pluspd_rxdma_get_tableid_from_tablename (const char *tablename)
 {
     for (uint32_t i = 0; i < p4pd_rxdma_tableid_max_get(); i++) {
